vision-explorer2/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
anthropic>=0.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
